+++ src/components/masters/sap-masters/part-master/AssemblyPartListing.js
@@ -140,10 +140,11 @@ class AssemblyPartListing extends Component {
     * @description Renders buttons
     */
     buttonFormatter = (cell, row, enumObject, rowIndex) => {
+        const { EditAccessibility, DeleteAccessibility } = this.props;
         return (
             <>
-                <button className="Edit mr5" type={'button'} onClick={() => this.editItemDetails(cell)} />
-                <button className="Delete" type={'button'} onClick={() => this.deleteItem(cell)} />
+                {EditAccessibility && <button className="Edit mr5" type={'button'} onClick={() => this.editItemDetails(cell)} />}
+                {DeleteAccessibility && <button className="Delete" type={'button'} onClick={() => this.deleteItem(cell)} />}
             </>
         )
     }
@@ -231,7 +232,8 @@ class AssemblyPartListing extends Component {
     * @description Renders the component
     */
     render() {
-        const { isEditFlag, PartId, isOpenVisualDrawer, } = this.state;
+        const { isOpenVisualDrawer, } = this.state;
+        const { AddAccessibility, } = this.props;
         const options = {
             clearSearch: true,
             noDataText: <NoContentFound title={CONSTANT.EMPTY_DATA} />,
@@ -253,13 +255,11 @@ class AssemblyPartListing extends Component {
                     <Col md="4" className="search-user-block">
                         <div className="d-flex justify-content-end bd-highlight w100">
                             <div>
-
-                                <button
+                                {AddAccessibility && <button
                                     type="button"
                                     className={'user-btn'}
                                     onClick={this.displayForm}>
-                                    <div className={'plus'}></div>ADD BOM</button>
-
+                                    <div className={'plus'}></div>ADD BOM</button>}
                             </div>
                         </div>
                     </Col>
